(PC-37957)[API] refactor: move `cds_serializers` in `external_bookings`
