+++ api/src/pcapi/core/external_bookings/cds/serializers.py
