+++ api/tests/routes/pro/post_venue_provider_test.py
@@ -6,10 +6,10 @@ import pytest
 import pcapi.core.offerers.factories as offerers_factories
 import pcapi.core.providers.factories as providers_factories
 import pcapi.core.providers.repository as providers_repository
-from pcapi.connectors.serialization.cine_digital_service_serializers import IdObjectCDS
-from pcapi.connectors.serialization.cine_digital_service_serializers import ShowCDS
-from pcapi.connectors.serialization.cine_digital_service_serializers import ShowTariffCDS
-from pcapi.connectors.serialization.cine_digital_service_serializers import ShowsMediaoptionsCDS
+from pcapi.core.external_bookings.cds.serializers import IdObjectCDS
+from pcapi.core.external_bookings.cds.serializers import ShowCDS
+from pcapi.core.external_bookings.cds.serializers import ShowTariffCDS
+from pcapi.core.external_bookings.cds.serializers import ShowsMediaoptionsCDS
 from pcapi.core.external_bookings.models import Movie
 from pcapi.core.history import models as history_models
 from pcapi.core.providers.factories import CinemaProviderPivotFactory
